fix(tests): avoid duplicate precheck module names
